+++ code/core/src/core-server/change-detection/GitDiffProvider.ts
